requirements.txt
python-dotenv
django
djangorestframework
mysqlclient
numpy
faiss
torch
transformers
sentence-transformers
celery
redis
